+++ 18Lights/vklearnin/src/rendering/RenderConfig.cpp
@@ -14,8 +14,8 @@ int32_t  RenderConfig::window_pos_x  = 0;
 int32_t  RenderConfig::window_pos_y  = 0;
 float    RenderConfig::window_aspect = 0.0f;
 
-uint8_t RenderConfig::image_count = 2u;
-bool    RenderConfig::vsync_on = true;
+uint32_t RenderConfig::image_count = 2u;
+bool     RenderConfig::vsync_on    = true;
 
 uint8_t RenderConfig::msaa_samples = 1u;
 float   RenderConfig::anisotropy   = 1.0f;
